fix: fixed circular dependency giving errors under webpack
